Converted polkadex_balance_storage into a module
