+++ enclave/src/polkadex_balance_storage/balances.rs
@@ -1,0 +1,15 @@
+use codec::{Decode, Encode};
+use polkadex_sgx_primitives::Balance;
+
+
+#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq)]
+pub struct Balances {
+    pub free: Balance,
+    pub reserved: Balance,
+}
+
+impl Balances {
+    pub fn from(free: Balance, reserved: Balance) -> Self {
+        Self { free, reserved }
+    }
+}
