+++ enclave/src/polkadex_balance_storage/polkadex_balance_key.rs
@@ -1,0 +1,18 @@
+use codec::{Decode, Encode};
+use polkadex_sgx_primitives::{AccountId, AssetId};
+
+
+#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq)]
+pub struct PolkadexBalanceKey {
+    pub asset_id: AssetId,
+    pub account_id: AccountId,
+}
+
+impl PolkadexBalanceKey {
+    pub fn from(asset_id: AssetId, account_id: AccountId) -> Self {
+        Self {
+            asset_id,
+            account_id,
+        }
+    }
+}
